move MilliscondFormatter to its owned file  - update module exports and prelude
