+++ src/prelude.rs
@@ -1,5 +1,6 @@
 pub use crate::pretty::parser::*;
 pub use crate::pretty::splitter::*;
+pub use crate::pretty::MillisecondFormatter;
 pub use crate::relative::formatter::*;
 pub use crate::relative::parser::*;
 pub use crate::relative::*;
