+++ src/pretty/mod.rs
@@ -1,3 +1,9 @@
+mod formatter;
 pub mod parser;
 pub mod splitter;
 pub mod utils;
+
+pub use formatter::*;
+pub use parser::*;
+pub use splitter::*;
+pub use utils::*;
